Frontend Status: Add track data.
Provides a list of audio and subtitle tracks that are available in the
current stream. The list is a set of key/value pairs where the key is
the action that, if sent to the frontend, will select the given track
and the value is the translated description. As for chapter details, the
total number of audio and subtitle tracks is indicated in the main
'FrontendStatus' section and non-zero values indicate the presence of
additional 'SubtitleTracks' and 'AudioTracks' sections.

This is missing external subtitles support and an indication of the
current, in use track.

Rather than add an indeterminate number of, for example,
SELECTSUBTITLE_20 actions (some blu ray discs have over 20 audio tracks)
or refactor the event handling in tv_play, I've added an extra filter to
allow through certain 'selector' actions.