Fixes #10172. Cleanup duration calculations.
We just set the 708 duration on clears as if we were getting a regular caption update + we introduce some sanity checking so if length is not initialized or initialized with a 0 ms duration we leave the caption on screen for 750 ms.

Jim, please open new tickets for any additonal problems you notice.