Respect config.xml and mysql.txt files if they exist.
At some point LoadDatabaseSettings() lost it's return value (it was still documented though)
and we started using UPnP to connect to a random mythtv database whenever we were unable to
connect to the configured database.

This returns the return value for mysql.txt parsing and does not do autodetection
if there is a mysql.txt. This also adds similar parsing of config.xml and also
respects those db connection values if they exist. If those files exist but we
are unable to connect to the configured database we bring up manual config in
GUI apps and abort startup of mythbackend. If neither of those files exist we use
UPnP autodetection.