Seek backward optimization.
This improves the existing short seek optimization in FileRingBuffer, when
seeking backward we know we are no closer to the end of the buffer than we
were previously, so if reads were allowed they should still be allowed.

It may be possible to further optimize this by actually setting readsallowed
to true on backward seeks however that doesn't appear necessary and would
complicate the code by having multiple places where readsallowed is set to
true (currently this is only done in the main read ahead loop.)

This hot-point as well as the original solution were discovered by Chris Pinkham.