Allow reopening a writeable ringbuffer with a new filename.
This functionality flushes the original file, closes it, and
reopens the file or opens a new file if an optional new filename
is supplied.

Adds ReOpen(QString newFilename) methods to RemoteFile,
ThreadedFileWriter, RingBuffer, and FileRingBuffer.  A 'REOPEN'
command is also added to the remote file transfer protocol.

This functionality will be used in a soon-to-be-committed feature.

NOTE: The binary API version changes with this commit, so
      make clean, etc..

NOTE2: This also bumps the protocol version to 70 because of
       the new file transfer command, so all components need to
       be upgraded at the same time.