Add "tuningTimeout" hidden setting.
I know hidden settings are unpopular, but I doubt there are many people outside
devs who are really going to muck with this setting. If I'm wrong I'll gladly
create a UI for it. All it does is allow you to fiddle with the timeout for
giving up on a recording when there is a signal problem (currently hard coded
at 180 seconds.)