Services: *** Make Services require UTC DateTime ***
PLEASE READ!

As MythTV will use UTC time internally in the near future, to
avoid disrupting initial adopters of the Services API,
adjust input and output of the API to require UTC.

What does this mean to current API users?  It means when
presenting the date to the user, you will need to translate
into local time.  It means the API will output times and
dates in UTC format, and you will need to input times and dates
in UTC format.  If you are getting all your data from the API,
then plugging that same data back in to other functions, your
application should require no modification aside from user-facing
strings.

What does it mean to MythTV devs?  It means that when Myth goes
to internal UTC datetimes, the code below will need to be partially
reverted, to remove the "toUTC()"s on output, and the "toLocalTime()"s
on input.  The user facing strings (HTML setup) should remain the same.