Content API: Add a GetFilteredLiveStreamList API.
This is a simple API that allows narrowing the list of available
live streams by filename.  In a sense, it can ask as an "exists"
for streams made from an input file.  If a client app produces
multiple streams at multiple resolutions, this API would allow
one to return a list only of those streams, rather than having
to parse the list of all available streams.

Usage:

http://BackendServerIP:6544/Content/GetFilteredLiveStreamList?FileName=foo.ext