Fix video duration for videos without a seektable.
3924df8 removed an extra call to av_estimate_timings() for playback
startup optimization, but this call appears necessary to set the
duration of videos without a seektable in the DB.  There may be
a proper fix inside avformatdecoder.cpp or the ffmpeg code,
but I haven't figured out one yet.  This commit adds back the
call to av_estimate_timings() in another location in the code so
that it only executes if we were unable to determine the video
duration by any other methods.  This should allow faster startup as
intended in 3924df8 but also allowing for valid duration detection
for files without a seektable in the database.