MythPlayer: Add some high level playback flags.
This is the first step in consolidating a number of currently disparate
variables that control the high level functionality of MythPlayer and
its sub-classes.

The aim here is to make initialisation a little more explicit and then
allow more flexibility around the choice of video output type and
decoder preferences.

With this commit, it is now possible to explicitly allow the use of
certain hardware based video decoders for transcoding, commflagging,
preview generation etc by passing in KDecodeAllowEXT to the player
constructor. This will enable VDA and CrystalHD decoding if available -
both of which are well suited to these types of operation as the lack of
interlacing field detection may not be an issue, the real time playback
issues that both can suffer from shouldn't be relevant and both return
decoded frames to main memory - not video memory.