Update some strings in OSD for consistency with UI.
Changes

No, keep it, I changed my mind -> No, keep it

as was done for Watch Recordings in f7ef39e9fd3

Changes:

Delete this recording? -> Are you sure you want to delete:

as it appears in the (non-OSD) UI in Watch/Delete Recordings screens.
This removes the last occurrence of "Delete this recording?", but leaves
"Delete this recording" (as a proposed action, not a question, in the
end of playback prompt).

Refs #9227.

Signed-off-by: Michael T. Dean <mdean@mythtv.org>