Remove the HOWTO, which is now in the wiki
The MythTV HOWTO has been moved to the wiki (
http://www.mythtv.org/wiki/MythTV-HOWTO ) to make it easier for more
users to help maintain it.  Distributable "offline" copies of the HOWTO
will be made from the wiki pages in various formats and be made
available with releases.

More information will be provided in the future regarding the plan for
editing and updating the HOWTO, how to deal with changes that are
version specific, as well as what's expected of users updating the
pages, and any licensing issues/requirements.  Any substantive
changes--especially changes that change the meaning/interpretation of
MythTV-specific information (as opposed to distribution-specific
information) presented in the HOWTO--should generally be discussed on
the "discussion" page associated with the MythTV-HOWTO page, first.

I also removed the docs/i18n directory, which contained only one file,
keys_de.txt, which was a German-language translation of the keys.txt
file, since the file has not been updated since it was added on June 22,
2005, and much has changed in the keys.txt file, since.  Instead, any
translations of the keys.txt file (or any other part of the HOWTO)
should be done as separate wiki pages and can be linked from the HOWTO.

Makefile update will come separately, soon.