Add initial README.HOWTO to docs
Add a simple README.HOWTO to the docs directory, to direct people to the
wiki version of the HOWTO.  Anyone with better ideas for what it should
say should feel free to update it however they like.  This can be
removed, if appropriate, if we decide to include offline copies of the
HOWTO in the repo (as opposed to just in/with release tarballs).