Adds 'damaged' video property for poor recordings.
This sets the 'warning' state for the Watch Recordings list item if the
property is set and also adds a 'videoquality' statetypes for those themes
that wish to use an icon for this property.

The idea is that the recording is perhaps watchable, but this gives you
a warning that a there may be a serious problem with the recording so
you should perhaps check it before inviting a friend over to watch it
with you.

The property is set whenever RecordingQuality::IsDamaged() returns true.
This is primarily based gaps in the recording with extra weight given
to gaps at the beginning and end of the scheduled program slot. In the
future we may add transport stream discontinuity or other recorder
specific quality metrics.