Don't resync transport stream on every transport error...
If the FEC fails then the driver/hardware will set the transport error bit in our TS packet, we
were resyncing every time this happens, but we don't have to. This caused unnecessary work to be
done and resulted in some extra debugging being printed by the HDHomeRun stream handler.