VDPAU: Add a Null VDPAU renderer class for assisting transcode etc.
N.B. This is lightly tested and as yet not fully hooked up.

To test you will need to ensure the backend is run from a terminal to
give it access to an X session and then pass kDecodeAllowGPU to the
player constructor.

Once enabled and in use, VDPAU will be used to decode the video stream
and decoded frames are then passed back to main memory for subsequent
processing. Frames are currently passed back unscaled. Scaling may be
added as an option but it requires a conversion to RGBA in the GPU.
Hence there will be a trade off between the size of the frame data
passed back to main memory (the RGBA frame may be smaller but requires a
greater BPP) and the need to convert the frame back to YV12 in the CPU.

In testing, multiple streams can be decoded simultaneously on the same
GPU (and while using the same GPU for normal playback) but performance
will obviously be constrained by the capabilities of the video decoder
and the amount of GPU memory available.