Make MythSystem not crash when rudely deleted
Seems that there are some places (channel changing in particular) that like to
delete the MythSystem instance before we are really finished with it.  When
this happens, and then the child exits, MythSystem crashes.  This should fix
the issue by logging that the structure was deleted, and carrying on.

Fixes #10227