Fix expr regexp portability in configure gxx_version check.
Remove a redundant '^' from expr regexp in configure.

Per expr documentation, an anchored pattern match takes 2 arguments such
that, "The arguments are converted to strings and the second is
considered to be a (basic, a la GNU 'grep') regular expression, with a
'^' implicitly prepended."

See:  info --file coreutils.info --node 'String expressions'

Though the redundant '^' does not cause, at least, GNU expr to fail,
including it may be a problem on some expr implementations and causes
some versions of expr to complain:

expr: warning: unportable BRE: `^4\\.[3-9]': using `^' as the first
character

on stderr.