Add support for Side by Side and Top and Bottom 3D.
- add a new OSD 3D menu option (under Video).
- user can then select to discard one stereo field (right or top) for
non-3D capable TV sets or select Side by Side or Top and Bottom to
indicate that this is a 3D stream that will be shown in 3D by the
display.
- when Side by Side or Top and Bottom is selected, all on screen
graphics are rendered into each stereoscopic field at half width or
height. This includes main UI widgets when embedded (e.g. program
guide), visualisers, OSD and PiP/PBP. This is entirely untested on an
actual 3D capable set but should work as expected. There is no disparity
adjustment between the 2 overlays, so the OSD should appear in the plane
of the TV (which may need adjusting).

N.B. Currently only supported when using OpenGL rendering. VDPAU and
Direct3D support will follow.

Additional support for auto-detecting stereoscopic streams is also on
the way.