Do not keep trying to mount when it fails
Refs #10221

If a USB stick (or other removable drive) is attached to the frontend, and the
user running the frontend is not able to mount it with "mount /dev/sdc1" or the
like, we do not want to keep trying (twice a second!) to mount it.  There's
no purpose to that other than extra load, and filling log files with useless
crap.