Change media monitoring interval from 500ms to 5s
It seems excessive to hit the CDROM drive every 500ms to see if media is
there or not.  Every 5s should do fine.  (Also moved the interval to a
macro (#define MONITOR_INTERVAL) to make it more maintainable).

Ideally, we should try to use dbus for this if possible (where supported)
and not poll at all.