Make the ThumbGenerator thread actually abort
When you exit mythgallery, it deletes the ThumbGenerator, which causes it to
wait for completion.  As the inner code was never checking for an abort,
sometimes this would take a very long time, especially when dealing with
RAW format pictures.

Now, when cancel() is called, a flag is set that is checked in the inner
loops to cause a more immediate exit.

Fixes #10185