In progress TMDB v3 grabber script.
The metaclass and decorators in util.py are used to construct a
just-in-time interface to the new TMDB API based off the service
definitions in tmdb_api.py.  tmdb3.py is a minimally functional grabber
based off this interface, able to perform seaches and pull specific
movie searches, using the VideoMetadata class from the bindings to
construct the XML to send to MythTV.

The metaclass and decorators are messy and need to be cleaned up and
reorganizes.  The upstream API, as well as this interface to it, are
both incomplete, and not yet ready for use.