Prevent premature exit from DRB.
Now that we track recording gaps we really shouldn't abandon a recording because
the signal is out for a few minutes. For the HD-PVR we need to reset the device
when we don't see data for a while, but for DVB recorders we just need to keep
trying on the existing open file descriptor. So this adds a variable to the DRB
called poll_timeout_is_error, which is set to true for the HD-PVR and false for
DVB devices. When set we error out on a 2.5 second timeout so the device can be
reset, when clear we just continue waiting until we see some data or the DRB is
told to stop recording.