MythMusic: Fix a crash when switching visualizers
Switching from Piano to Spectrum or Squares was causing various random crashes
caused by the FFTW lib using memory it shouldn't be using. The route cause is
the different sampling sizes used by Piano compared to the others.

This just sanity checks the VisualNodes samples length is sane before using it
and also (not really required but doen't hurt anything) remove all VisualNodes
for a fresh start when switching visualisers.