Allow easier control of card input ordering.
This change adds two new settings at the card input level.  These
settings allow fine-grained control of the order inputs are used for
scheduling and live TV without the need to delete and recreate the
cards and inputs in the desired order.

"Schedule Order" defines the order inputs are used when scheduling
recordings and priorities are equal.  The available input with the
lowest, positive order number is used first.  Previously, the input
with the lowest input ID was used first.  If the schedule orders for
two inputs are equal, the old behavior is used as a fallback.  Setting
the schedule order for an input to zero makes that input unavailable
for scheduling.

"Live TV Order" defines the order inputs are used when entering live
TV.  The available input with the lowest, positive order number is
used first.  Previously, the default input on the card with the lowest
(or highest if "Avoid Conflicts..." was set) card ID was used first.
If the live TV order for two inputs are equal, the order they are used
is undefined.  Setting the live TV order for an input to zero makes
that input unavailable for live TV.  This setting obsoletes the old
"Avoid conflicts between Live TV and scheduled shows" setting.

Setting either schedule order or live TV order to zero for an input
effectively restricts the use of that input to the other activity.  In
other words, it is now possible to dedicate an input exclusively to
scheduling only or live TV only.

Setting both schedule order and live TV order to zero effectively
takes that input off-line as far as MythTV is concerned.  That input
can then be safely used for other purposes outside of MythTV as long
as the user avoids conflicts with other inputs on the same card or in
the same input group.

Note: this change doesn't yet do exactly what it claims.  The input
used when entering live TV is actually the default input on the card
with the input having the lowest, positive live TV order.  This will
be corrected in a later commit.