Finish input ordering changes started in commit 3281cdd3.
Honor the livetvorder setting when choosing the starting input for
live TV.  Onc correction from the previous commit is needed.  It said
that the available input with the lowest, non-zero livetvorder is used
first.  It is actually the available, local input witht the lowest,
non-zero livetvorder.  If no local input is available, it is the
available, remote input with the loweset, non-zero order.  In other
words, available inputs on local recorders are preferred over remote
inputs.

Please note that this change also does away with the concept of the
"defaultinput" for a card.  It was actually the last used input.  It
is no longer needed since livetvorder determines the chosen input.