Fix warnings in 1262 DB upgrade.
Fix warnings that can occur in the 1262 DB upgrade if any of the columns
being changed to varchar contain NULL values.

These warnings resulted in a DB upgrade failure for a at least one user,
but in attempting to reproduce the failure (with Qt4.6.2+MySQL5.1,
Qt4.6.2+MySQL5.5, Qt4.7.3+MySQL5.1, Qt4.8.0+MySQL5.1, and
Qt4.8.0+MySQL5.5), I was unsuccessful.  The only way I was able to cause
it to fail was with MySQL strict mode enabled
(sql_mode=STRICT_ALL_TABLES).  Note that many other parts of MythTV are
likely to fail with strict mode enabled, so it is recommended that users
do not enable strict mode.

Thanks to dchilton for reporting the issue, and to JK for helping to
troubleshoot.