Add a simple Bonjour/Avahi service registration class.
This should work out of the box on OS X but requires the Avahi Bonjour
compatability layer on *nix (libdns_sd).

Fairly straightforward service registration and removal that will be
used by a couple of upcoming features (and could easily be used as an
extra layer of 'advertising' by both mythfrontend and mythbackend).