Mythfrontend: Add Remote Audio Output Protocol streaming support.
Allows streaming of audio from iOS devices to mythfrontend.

Seems to be working pretty well, though there may be occasional av sync
issues when streaming audio from a device that is playing the video.
Frontends will appear as MythTV on TheNameofYourFrontend in apps such as
iPod, YouTube etc.

The RAOP server will try to autodetect a suitable port to listen on,
individual connections will search for a suitable data port to open,
devices not using the default port will adjust their name accordingly
and multiple connections can be open on each frontend at a time (and
will all be played if your configured audio device supports it)- all of
which means you can have multiple RAOP configured frontends on the same
machine each playing back multiple streams...

There is no configuration but I will add a setup page (for this and
other interfaces) that will allow it to be disabled per frontend.

It should also be possible to tie into mythmusic to allow for
visualisations.

Reguires both Bonjour support (for service announcement) and libcrypto
for decoding the audio.