MythUIText:  Fix a fringe case when a single line of text is *just* the wrong size, and is being cutdown.
QFontMetrics seems to think more 'text' fits in a given width, than
QTextLine.  I don't know how to determine exactly what the difference is, so
tell QFontMetrics that the width is averageCharWidth() less than it really
is.

Also force the text to be rendered, even when QTextLine thinks it doesn't
fit, in case averageCharWidth() is too small.  This may result in not all of
the '...' from being rendered, but it is better than rendering no text at
all.