Rework pipe handling in System class
This adds a DequeBuffer class, as a dynamically expanding buffer for use
with POSIX pipes. The class includes a shared poller thread that will
continually push or pull data to the respective pipe to prevent the
joined application from blocking.

The System class now uses this as temporary storage for stderr and
stdout, to get around the previous issue of applications blocking once
they hit the 64KB pipe limit.