MythUIText: Allocate enough "space" for any shadows and outlines when rendering the text.
This fixes a bug, where shadows and outlines where being clipped by the
"viewport".

Note that the behavior for outlines is a little different than it was
previous to the conversion to QTextLayout.  The old system would move the
text down and to the right, to make room for the outline.  The problem with
that, is that "centered" text was no longer centered.  The new code will
leave the text position alone, and add space all the way around it, for the
outline.  I don't know of any themes which actually use outlines, so I don't
expect this to break anything.