Correct 32-bit false positive autoexpiration when free space is >2TB.
This corrects the symtomatic instances of improper typing in the
autoexpiration code.  The total free space (in KB) is cast from a
64-bit+ long long, to a size_t which varies depending on architecture.
When compiled on a 32-bit system, this unsafe casting results in an
inversion of the conditional every 2TB of freespace.  At 0-2TB free, or
4-6TB free, everything works properly.  At 2-4TB or 6-8TB free, it
immediately deletes anything it can.