MythMusic: Refactor the music start directory handling
We was passing and storing the music start directory in several places this
changes everywhere to use the one in gMusicData.

Also change the metadata class to store the tracks path and filename minus the
start directory. This is one tiny step closer to supporting storage groups and
has the added benefit of saving approx. 200K of memory on my setup.