RAOP: Better handling of multiple audio streams.
If the audio device cannot be opened, assume something else is already using it and the device cannot be shared (e.g. spdif, hdmi etc). In this case, retry the audio device every 5 seconds. If multiple RAOP connections are open, only allow one to retry at a time.

When the audio device is available again (e.g. after watching tv), the next RAOP connection in the queue should start playing.