Add QUERY_ACTIVE_BACKENDS call to backend protocol.
This adds a new call to the backend protocol to enumerate a list of all
currently available backends, intended for use when scanning for
non-recorded content on multiple backends. This allows offline backends
to be skipped outright, rather than having to attempt connection and
wait for a timeout.

This bumps the protocol version to 72.