Rework "Minor scheduler cleanup and correction."
This commit reworks commit 65b63e1f to avoid the performance
regression.  It's technically wrong, since it doesn't consider two
programs with different titles, but the same programids as being the
same.  The performance drop is worse than trying to handle this
exceedingly rare case which the previous code didn't handle either.