MythMusic: Use the portable libcdio for reading CDs
This patch enables MythMusic to play and rip CDs using libcdio on all
platforms and also adds code to replace libcdaudio CDDB lookups.

NOTE: This removes libcdaudio and the CDDA Paranoia libraries as dependencies
and adds libcdio as an optional dependency (if it's missing you wont be able
to play or rip CDs). Closes #9362, #7857.

NOTE: It seems the freedb.freedb.org server is currently down so I had to do
some of the testing using the musicbrainz mirror. It was working fine yesterday
so I've left it hardcoded to the freedb server for now.