Initial AirPlay support.
This is currently disabled by default and is enabled by setting the
temporary environment variable 'MYTHTV_AIRPLAY'.

Outstanding issues:-

- playback will silently fail if there is already a video playing on the
frontend. This usually encourages the source device to blacklist the
frontend for an indeterminate period of time.
- playback exit isn't properly handled. Currently the player is
restarted when told to stop and exits gracefully when no more data is
available. This does not work for remote sources (e.g. YouTube - where
the data is provided directly from the remote site to the frontend).
- no seeking/scrubbing with remote sources such as YouTube.
- initial seek/bookmark positioning does not work.
- scrubbing does not work on devices with iOS 5 (at least I think it's
iOS 5 related).
- playback fails with some remote services (e.g. BBC iPlayer). This is a
problem in the StreamingRingBuffer code.
- the playback url is sometimes provided embedded in a binary property
list (e.g. YouTube). Parsing of this plist is currently only supported
on OS X.

All that said, when it works, it works well - with all the features of a
full frontend (e.g. timestretch)