mythtranscode: --fifoinfo reports aspect ratio of first uncut frame
This fixes a problem using the --fifoinfo option to report the format
of a recording.  When using a cutlist the first few frames may be cut
and the format of the first uncut frame may be different from the first
physical frame.

This fixes a problem when using nuvexport etc if the recording starts
in widescreen e.g. for adverts but switches to 4:3 for the feature.

Fixes #10301

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Gavin Hurlbut <ghurlbut@mythtv.org>