This permits the translation of additionnal strings in the status screen and also change the context of others so that they are regrouped in the same context as the other status screen strings.
The problem reported in ticket #10251 was not actually caused by
MythTV's code which actually handled the translation of those strings
correctly, the problem is that Qt's tools were not extracting
those strings when they should have. I have modified our code to
work around the problem.

Sorry for the delayed commit... Figuring out the problem was easy but I
couldn't test my fix because of defective hardware which has now been
replaced.

Closes #10251