Improve scheduler handling of seriesids and programid authorities.
Programs in the same series occasionally have different titles.
Previously, the scheduler would not match programs with titles
different from that of the recording rule.  The scheduler now uses
seriesid in addition to title to find matching programs.  The various
screens such as the EPG which offer to list "Other showings" now also
match on seriesid as well as title.

There is now a "This series" filter to aid in scheduling when two
series share the same title.  If this filter is enabled for a
recording rule, the scheduler only matches programs for that series
and does not match programs from other series even if they have the
same title.  This change also corrects the description of a couple of
other filters.

Some users have video sources with programids generated from different
authorities.  Previously, if two programs had non-empty programids,
the scheduler considered them different programs even if the programid
authorities were different.  The scheduler now falls back to
title/subtitle/description checking if two programs have non-empty
programids from different authorities.

The additional, programid authority checking is less efficient than
the previous, simple, programid checking and results in longer
scheduling times.  To remedy this for users who don't need it, the
scheduler checks the database at initialization and only performs the
authority checking if two or more distinct authorities are detected.