DVR Service: Make GetRecordScheduleList work on an MBE.
Somehow I only tested this on my slave, development, BE.  Existing
method would not work on an MBE as the master cannot connect to
itself.

This makes a formerly private method in the scheduler public,
but this seems relatively safe as it is only the DB query from
the record table, and the subsequent filling of a RecInfo List.