Show alternate names for mythlink.pl arguments
The wiki page used the --link argument for examples, but it was not
shown in --help output.  This just updates the help output to show
alternative names for the argument.

Also remove the redundant "dest" name for "destination" (as "dest" is a
sufficiently unique portion of "destination" to qualify as the
destination argument) and "h" for "help".

Thanks to Xipmix for noticing the discrepancy in the wiki page.