Allow specifying a maximum length in mythlink.pl.
This provides a way for users who use inappropriate data to construct a
link name (such as a description) to specify a maximum length of the
link name to work around limitations imposed by the file system.  Note
that the names will be truncated and a count variable added to them to
keep within the --maxlength specified.  The names aren't necessarily
useful or descriptive, but the feature was requested and was easy enough
to add--and may actually convince those users to choose more approrpiate
data for use in creating file names.