Services: Add a AddRecordSchedule API.
This POST-only API will add (and technically speaking, modify where a pre-existing
rule exists) a recording rule for a given chanid and starttime.  Because most
of the Services API is build around the chanid and starttime info, and because I
didn't want to make an already ungodly-long list of arguments even longer, I've
opted for this first method to support only those people who actually have listings
information.  In other words, this API cannot be used to create manual recordings,
though that can be added as a separate method (and I suppose this one could be
retrofit to use it like the image content APIs all use GetImageFile).

This API takes all the recording-specific arguments in addition to chanid and
starttime, but at its simplest, one can create a recording rule as follows:

http://BackendIP:6544/Dvr/AddRecordSchedule?ChanId=2066&StartTime=2012-02-07T10:00:00

This will create a single recording rule, matching duplicates against all,
with a matching method of subtitle and description.  It will use the Default
Storage Group, Playback Group, and Recording Profile.

A more complicated rule setup might be something like this:

http://BackendIP:6544/Dvr/AddRecordSchedule?ChanId=2066&StartTime=2012-02-07T10:00:00&Season=1&Episode=1&Inetref=12345&Type=Channel&DupIn=All&DupMethod=SubtitleThenDescription&RecProfile=LiveTV&MaxEpisodes=5&AutoMetaLookup=true&AutoExpire=false

Etc.  See code and wsdl for documentation.