Make Scheduler::GetAllScheduled() thread-safe.
Commit [fd582899] made a private function public so I went to check
if it was safe to do so. It wasn't but this was easily missed
because the public getAllScheduled() was already calling that
function without any locking for the recordTable variable access.

This adds locking to findAllScheduledPrograms() and then renames
getAllScheduled(QStringList&) to GetAllScheduled(QStringList&) and
findAllScheduledPrograms(RecList&) to GetAllScheduled(RecList&) for
naming consistency with the GetAllPending(*) methods. I also made
both functions const since they don't modify the state of the
Scheduled class.