Make Videos storage directory definitions global
This reworks the initial scan behavior for the Video library, making the
use of the Videos storage group consistent with those for recordings.
The hostnames specified for those groups are now ignored in effect,
allowing slave backends to use directories defined on the master
backend.

This is accomplished by adding a new method to supply a list of
available backends to query.  When the list of search paths are added in
later, the Myth URIs are extracted from the list, and collected in a
unique list of directories, in case the same directory is defined for
multiple backends.  All resultent directories are recombined with all
available backends, and added into the list.  Scan behavior for local
directories is left unchanged.  The failed list is replaced by an
offline list, which reports as a popup of inaccessible hosts with
existing video content referenced to them.

There is currently no mechanism to remove content lost to a failed
backend, besides setting up a replacement backend of the same name.
There is a flaw in the scan when performed by the master backend, where
it will not query remote backends for content lists. The current storage
group code will only write to directories defined locally for that
backend, meaning the various artwork storage groups still need to be
defined on each backend for that to work properly. I've held off
performing a schema update to move all of these definitions over to the
master backend until that artwork issue is resolved.