Update SCAN_VIDEOS call to provide host list following 1693a18
This supplies the active backend list directly to the video scanner,
rather than have it attempt to connect to the master backend to pull the
list, and fail because it's connecting back to itself. Note, this will
not have any effect until the previously mentioned issue of being unable
to access those slave backends to pull a filelist has been resolved.