Perform protocol validation against remotefile sockets.
This adds in a validation check to the RemoteFile socket creation code.
While validation is optional when connecting to the backend in
mainserver.cpp, allowing clients to ignore the check and connect
regardless, libmythprotoserver requires that step take place. As a
result, file transfers with mythmediaserver abort when it refuses
connection from the client.