Add ServerPool class.
This adds a new ServerPool class to libmythbase, with the purpose of
emulating the interfaces of both QTcpServer and QUdpSocket as is
currently required by the various interfaces MythTV supports. This
class replaces the previous use of QHostAddress::Any and
QHostAddress::AnyIPv6 to perform blanket binds, with a pool of
sockets each bound to individual addresses.

The following signals are available for use:
  newConnection(QTcpSocket *)
  newDatagram(QByteArray, QHostAddress, quint16 port)

Additionally, newTcpConnection(int) can be subclassed and overridden to
produce alternate socket types.