Add methods for IP selection.
This adds a mechanism to determine a list of addresses that servers
should listen to, based off those defined for the machine. If
BackendServerIP or BackendServerIP6 are defined for that host, only
those and localhost addresses are selected.  If otherwise, any
link-local addresses are filtered out, and IPv4 is limited to only
RFC1918 private ranges. If users wish to listen on an internet
accessible address, they will need to force that with those settings.

Also added is a GetBackendServerIP convenience function, replacing the
existing GetSettingOnHost calls, and returning an IPv4 or IPv6 address
as is available.