Split BackendServerIP into v4 and v6. Implement server pools.
This migrates all existing TCP and UDP servers over to the new
ServerPool class. This also adds a new BackendServerIP6 setting to the
table, and modifies all previous queries to that setting to handle the
fact that there are now two values. A schema update will automatically
set the existing address to the proper field, and populate the other
with the loopback address, effectively disabling access to anything but
the local machine on that protocol version.

Protocol queries, remote playback, webserver, UPnP, command socket, and
UDP listener for mythmessage have all been tested and should work. I do
not have any iOS devices to use with the RAOP code, so that is untested
beyond compiling.

Fixes #9820 Fixes #9840
Refs #10165 Refs #10288